bracket -> parentheses
It seems like most of the style guide suggests US standards, and in the US an unqualified "bracket" refers to [square brackets] rather than (parentheses or round brackets). I decided to change "brackets" to "parentheses" for simplicity. 

P.S. I would completely understand if this is extra pedantic and you prefer to disregard this PR :)